Convert main and stack_panel to Typescript
